요리 쪽 작업하기전 빽업

+++ Assets/Test/TestInventory/DataScript/Vars.cs
@@ -17,7 +17,7 @@ public static class Vars
     public static readonly int maxStamina = 100; // 회색 스태미나 절대로 변하지 않는 스태미너칸
     public static readonly float hunterMaxHp = 100;
     public static readonly float herbalistMaxHp = 100;
-    public static readonly int lanternMaxCount = 18;
+    public static readonly float lanternMaxCount = 18;
 
     // Localization
     public static Localization localization = Localization.Korean;
